requirements.txt
python-vlc>=3.0.20
fastapi>=0.115.0
uvicorn>=0.32.0
yt-dlp>=2026.3.0
pytubefix>=8.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
midi2audio>=0.1.1
spotipy>=2.23.0
tinytag>=2.0.0
websockets>=15.0
PySide6>=6.6
